project restructured into modules [Fixes #159828947]
