+++ requirements.txt
@@ -13,6 +13,7 @@ coveralls==1.3.0
 docopt==0.6.2
 Flask==1.0.2
 Flask-RESTful==0.3.6
+gunicorn==19.9.0
 idna==2.7
 isort==4.3.4
 itsdangerous==0.24
